nz-lightrag/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
lightrag==0.1.0b6
psycopg2-binary==2.9.9
qdrant-client==1.9.1
openai==1.25.0
sqlalchemy
pgvector
httpx==0.27.0